+++ haskell-lsp-types/src/Language/Haskell/LSP/Types.hs
@@ -26,7 +26,6 @@ import           Language.Haskell.LSP.Types.CodeAction
 import           Language.Haskell.LSP.Types.Color
 import           Language.Haskell.LSP.Types.Command
 import           Language.Haskell.LSP.Types.Completion
-import           Language.Haskell.LSP.Types.Constants
 import           Language.Haskell.LSP.Types.Diagnostic
 import           Language.Haskell.LSP.Types.DocumentFilter
 import           Language.Haskell.LSP.Types.FoldingRange